+++ libsolidity/ast/ASTJsonImporter.cpp
@@ -552,6 +552,7 @@ ASTPointer<FunctionDefinition> ASTJsonImporter::createFunctionDefinition(Json::V
 		freeFunction,
 		kind,
 		memberAsBool(_node, "virtual"),
+		memberAsBool(_node, "suffix"),
 		_node["overrides"].isNull() ? nullptr : createOverrideSpecifier(member(_node, "overrides")),
 		_node["documentation"].isNull() ? nullptr : createDocumentation(member(_node, "documentation")),
 		createParameterList(member(_node, "parameters")),